airflow/requirements.txt
# RSS and Web Scraping
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# Google Cloud Storage
google-cloud-storage==2.14.0

# Vector Database
chromadb==0.4.22

# Ollama Client
ollama==0.4.4

# Text Processing
langchain-text-splitters==0.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
email-validator>=2.0.0
